requirements.txt
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
google-api-core==2.8.0
google-auth==2.6.6
google-cloud-bigtable==2.9.0
google-cloud-core==2.3.0
googleapis-common-protos==1.56.1
grpc-google-iam-v1==0.12.4
grpcio==1.46.3
grpcio-status==1.46.3
idna==3.3
peewee==3.14.10
proto-plus==1.20.4
protobuf==3.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.27.1
rsa==4.8
six==1.16.0
typer==0.4.1
urllib3==1.26.9
